Change non functional molecule name
As ansible-nfv repository starts to integrate molecule testing for the
roles, the gate will look for the molecule directory within the roles in
order to execute it.
Since molecule in "tuned_configuration" is non functional and serve more
as an example, changing the name to molecule_example.
Later this role will be provided with functional molecule testing.

Change-Id: If079ee6f6c861ad71b751dd0ed63f3fbaadd04ae